src/observability/pyproject.toml
[project]
name = "mitol-django-observability"
version = "2026.1.0"
description = "MIT Open Learning Django plugin for OpenTelemetry, structured logging, and alerting"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.10"
dependencies = [
    "django>=4.2",
    "mitol-django-common",
    "opentelemetry-api>=1.31.0",
    "opentelemetry-sdk>=1.31.0",
    "opentelemetry-exporter-otlp-proto-http>=1.31.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.31.0",
    "structlog>=24.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
sentry = ["structlog-sentry>=2.0.0"]

[tool.bumpver]
current_version = "2026.1.0"
version_pattern = "YYYY.MM.DD[.INC0]"

[tool.bumpver.file_patterns]
"pyproject.toml" = ['version = "{version}"']
"mitol/observability/__init__.py" = ['__version__ = "{version}"']

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
include = ["CHANGELOG.md", "README.md", "py.typed", "**/*.py"]
exclude = ["BUILD", "pyproject.toml"]

[tool.hatch.build.targets.wheel]
include = ["CHANGELOG.md", "README.md", "py.typed", "**/*.py"]
exclude = ["BUILD", "pyproject.toml"]
